# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Create a Selfbeat AI comparison
 */
export const createSelfbeatComparisonBodyQuestionMax = 1000;

export const createSelfbeatComparisonBodyModeDefault = `live`;

export const CreateSelfbeatComparisonBody = zod.object({
  question: zod.string().min(1).max(createSelfbeatComparisonBodyQuestionMax),
  mode: zod
    .enum(["live", "mock"])
    .default(createSelfbeatComparisonBodyModeDefault),
});

export const CreateSelfbeatComparisonResponse = zod.object({
  id: zod.string(),
  question: zod.string(),
  timestamp: zod.number(),
  responses: zod.array(
    zod.object({
      model: zod.enum(["chatgpt", "claude", "gemini", "deepseek", "grok", "mistral", "llama", "perplexity", "cohere", "qwen"]),
      displayName: zod.string(),
      color: zod.string(),
      answer: zod.string(),
      selfCriticism: zod.string(),
      score: zod.number(),
      accuracyScore: zod.number(),
      selfAwarenessScore: zod.number(),
      status: zod.enum(["success", "fallback"]),
      error: zod.string().optional(),
    }),
  ),
  verdict: zod.string(),
  verdictDetails: zod.object({
    summary: zod.string(),
    bestAnswer: zod.string(),
    clearestAnswer: zod.string(),
    agreementPoints: zod.array(zod.string()),
    disagreementPoints: zod.array(zod.string()),
    overallWinner: zod.string(),
    explanation: zod.string(),
  }),
  isMedical: zod.boolean(),
  physicianNote: zod.string().optional(),
  source: zod.enum(["live", "mock", "mixed"]),
  cached: zod.boolean(),
  providerStatuses: zod.array(
    zod.object({
      model: zod.enum(["chatgpt", "claude", "gemini", "deepseek", "grok", "mistral", "llama", "perplexity", "cohere", "qwen"]),
      provider: zod.string(),
      status: zod.enum(["live", "fallback"]),
      message: zod.string(),
    }),
  ),
});

/**
 * @summary Get a cached Selfbeat comparison
 */
export const GetSelfbeatComparisonParams = zod.object({
  id: zod.coerce.string(),
});

export const GetSelfbeatComparisonResponse = zod.object({
  id: zod.string(),
  question: zod.string(),
  timestamp: zod.number(),
  responses: zod.array(
    zod.object({
      model: zod.enum(["chatgpt", "claude", "gemini", "deepseek", "grok", "mistral", "llama", "perplexity", "cohere", "qwen"]),
      displayName: zod.string(),
      color: zod.string(),
      answer: zod.string(),
      selfCriticism: zod.string(),
      score: zod.number(),
      accuracyScore: zod.number(),
      selfAwarenessScore: zod.number(),
      status: zod.enum(["success", "fallback"]),
      error: zod.string().optional(),
    }),
  ),
  verdict: zod.string(),
  verdictDetails: zod.object({
    summary: zod.string(),
    bestAnswer: zod.string(),
    clearestAnswer: zod.string(),
    agreementPoints: zod.array(zod.string()),
    disagreementPoints: zod.array(zod.string()),
    overallWinner: zod.string(),
    explanation: zod.string(),
  }),
  isMedical: zod.boolean(),
  physicianNote: zod.string().optional(),
  source: zod.enum(["live", "mock", "mixed"]),
  cached: zod.boolean(),
  providerStatuses: zod.array(
    zod.object({
      model: zod.enum(["chatgpt", "claude", "gemini", "deepseek", "grok", "mistral", "llama", "perplexity", "cohere", "qwen"]),
      provider: zod.string(),
      status: zod.enum(["live", "fallback"]),
      message: zod.string(),
    }),
  ),
});