# lib/api-zod/src/generated/types/selfbeatModelKey.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export type SelfbeatModelKey =
  (typeof SelfbeatModelKey)[keyof typeof SelfbeatModelKey];

export const SelfbeatModelKey = {
  chatgpt: "chatgpt",
  claude: "claude",
  gemini: "gemini",
  deepseek: "deepseek",
  grok: "grok",
  mistral: "mistral",
  llama: "llama",
  perplexity: "perplexity",
  cohere: "cohere",
  qwen: "qwen",
} as const;